Assert source of null arg throws is our lib
The operators listed below delegate some of their work to LINQ. They relied
on those LINQ method to do the actual null-argument checks and passed tests
on the basis that they shared the same parameter names (since the offending
parameter name was also asserted).

- ToDictionary
- ToLookup
- Index
- Partition
- OrderBy
- ThenBy

The NullArgumentTest is enhanced to check that the first frame from the
stack trace of a thrown ArgumentNullException came from our library.